+++ zephyrapp/game/src/cowtools/cat_math.c
@@ -98,8 +98,8 @@ CAT_vec2 CAT_vec2_rotate(CAT_vec2 a, float t)
 {
 	return (CAT_vec2)
 	{
-		cos(t) * a.x - sin(t) * a.y,
-		sin(t) * a.x + cos(t) * a.y
+		cosf(t) * a.x - sinf(t) * a.y,
+		sinf(t) * a.x + cosf(t) * a.y
 	};
 }
 
@@ -255,23 +255,23 @@ CAT_mat4 CAT_rotmat(float x, float y, float z)
 	CAT_mat4 X =
 	{
 		1, 0, 0, 0,
-		0, cos(x), -sin(x), 0,
-		0, sin(x), cos(x), 0,
+		0, cosf(x), -sinf(x), 0,
+		0, sinf(x), cosf(x), 0,
 		0, 0, 0, 1
 	};
 
 	CAT_mat4 Y =
 	{
-		cos(y), 0, sin(y), 0,
+		cosf(y), 0, sinf(y), 0,
 		0, 1, 0, 0,
-		-sin(y), 0, cos(y), 0,
+		-sinf(y), 0, cosf(y), 0,
 		0, 0, 0, 1
 	};
 
 	CAT_mat4 Z =
 	{
-		cos(z), -sin(z), 0, 0,
-		sin(z), cos(z), 0, 0,
+		cosf(z), -sinf(z), 0, 0,
+		sinf(z), cosf(z), 0, 0,
 		0, 0, 1, 0,
 		0, 0, 0, 1
 	};
